+++ app/Http/Controllers/TaskListController.php
@@ -46,6 +46,8 @@ class TaskListController extends Controller
      */
     public function show(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         return $this->success([
             'task_list' => new TaskListResource($taskList),
         ]);
@@ -56,6 +58,8 @@ class TaskListController extends Controller
      */
     public function update(Request $request, TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $request->validate([
             'title' => ['string', 'required']
         ]);
@@ -73,6 +77,8 @@ class TaskListController extends Controller
      */
     public function destroy(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $taskList->delete();
 
         return $this->success();
@@ -93,6 +99,8 @@ class TaskListController extends Controller
      */
     public function addTask(Request $request, TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $request->validate([
             'title' => ['required', 'string'],
             'details' => ['nullable', 'string'],
